Tests and config file (#38)
* Adding a tests (discovered a bug)

* Check for config file

* Remove extra lines

* Improved file name

* Minor output formatting (#20)

* Update

+++ R/cache.R
@@ -1,8 +1,6 @@
+get_cache_dir = function() tools::R_user_dir("oysteR", which = "cache")
 get_cache_file = function() {
-  dir = tools::R_user_dir("oysteR", which = "cache")
-  if (!file.exists(dir))
-    dir.create(dir, recursive = TRUE, showWarnings = FALSE)
-
+  dir = get_cache_dir()
   path = file.path(dir, "cached-deps.rds")
   return(path)
 }
@@ -11,6 +9,7 @@ ensure_cache = function() {
   path = get_cache_file()
   if (file.exists(path)) return(path)
 
+  dir.create(get_cache_dir(), recursive = TRUE, showWarnings = FALSE)
   audits = no_purls_case()
   audits$time = integer(0)
   class(audits$time) = c("POSIXct", "POSIXt")
@@ -56,8 +55,6 @@ update_cache = function(audits) {
 #' @export
 remove_cache = function() {
   path = get_cache_file()
-  if (file.exists(path)) {
-    file.remove(path)
-  }
-  return(invisible(NULL))
+  if (file.exists(path)) file.remove(path)
+  return(NULL)
 }
